+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - Edit History is now transparent and public
 - Fix Responsiveness on desktop
-
+- Save unpublished word to drafts
+- Load word from drafts
 
 ### 0.5.9+22 July 21, 2023